Рефакторинг, Корректировка
Фикс адаптера, создание событий, добавление гибкости, внедрения логгера во все сферы, файл конфигурации, класс кухни, система репутации, прочие мелкие изменения.

+++ Assets/Scripts/MyGame/UpgradeablePart.cs
@@ -1,16 +1,10 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace MyGame
 {
     public abstract class UpgradeablePart : Upgradeable
     {
         public UpgradeablePart() : this(0) { }
 
-        public UpgradeablePart(byte level) : base(level) { }
+        public UpgradeablePart(ulong level) : base(level) { }
 
         public abstract float GetModifier();
     }
